requirements.txt
# Core dependencies
openai>=1.0.0
azure-search-documents==11.4.0
azure-core==1.29.5
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1

# Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Logging and monitoring
colorlog==6.8.0

# Utilities
python-dateutil==2.8.2